Files: add cache reset, add blocking removing linkeg group

+++ common/ASC.Core.Common/Context/Impl/UserManager.cs
@@ -509,6 +509,12 @@ namespace ASC.Core
             if (Constants.LostUser.ID == userId || Constants.LostGroupInfo.ID == groupId) return;
             PermissionContext.DemandPermissions(Constants.Action_EditGroups);
 
+            if (CoreBaseSettings.VDR)
+            {
+                var group = GetGroupInfo(groupId);
+                if (group.CategoryID == Constants.LinkedGroupCategoryId) return;
+            } 
+
             UserService.RemoveUserGroupRef(Tenant.TenantId, userId, groupId, UserGroupRefType.Contains);
 
             ResetGroupCache(userId);
@@ -520,6 +526,8 @@ namespace ASC.Core
             PermissionContext.DemandPermissions(new GroupSecurityObject(groupId), Constants.Action_EditLinkedGroups);
 
             UserService.RemoveUserGroupRef(Tenant.TenantId, userId, groupId, UserGroupRefType.Contains);
+
+            ResetGroupCache(userId);
         }
 
         internal void ResetGroupCache(Guid userID)
